vetai-backend/requirements.txt
Flask==2.0.1
flask-cors==4.0.0
pandas==2.0.3
matplotlib==3.7.2
python-dotenv==1.0.0
openai==1.3.5 